+++ java/core/clustercontroller/src/test/java/io/joynr/messaging/LibJoynrMessageRouterTest.java
@@ -82,6 +82,8 @@ public class LibJoynrMessageRouterTest {
     private String unknownParticipantId = "unknownParticipantId";
     private Long sendMsgRetryIntervalMs = 10L;
     private int maxParallelSends = 10;
+    private long routingTableGracePeriodMs = 60000L;
+    private long routingTableCleanupIntervalMs = 60000L;
 
     private String globalAddress = "global-address";
 
@@ -104,6 +106,8 @@ public class LibJoynrMessageRouterTest {
                                                   provideMessageSchedulerThreadPoolExecutor(),
                                                   sendMsgRetryIntervalMs,
                                                   maxParallelSends,
+                                                  routingTableGracePeriodMs,
+                                                  routingTableCleanupIntervalMs,
                                                   messagingStubFactory,
                                                   messagingSkeletonFactory,
                                                   addressManager,
@@ -124,9 +128,13 @@ public class LibJoynrMessageRouterTest {
         messageRouter.route(message);
         Thread.sleep(100);
         final boolean isGloballyVisible = true;
+        final long expiryDateMs = Long.MAX_VALUE;
+        final boolean isSticky = false;
         Mockito.verify(routingTable).put(Mockito.eq(unknownParticipantId),
                                          Mockito.eq(parentAddress),
-                                         Mockito.eq(isGloballyVisible));
+                                         Mockito.eq(isGloballyVisible),
+                                         Mockito.eq(expiryDateMs),
+                                         Mockito.eq(isSticky));
     }
 
     @Test