+++ java/core/clustercontroller/src/test/java/io/joynr/messaging/routing/CcMessageRouterTest.java
@@ -114,6 +114,8 @@ public class CcMessageRouterTest {
             private Long msgRetryIntervalMs = 10L;
             private int maximumParallelSends = 1;
             private int maxMessagesInQueue = 10;
+            private long routingTableGracePeriodMs = 30000;
+            private long routingTableCleanupIntervalMs = 60000;
 
             @Override
             protected void configure() {
@@ -125,6 +127,11 @@ public class CcMessageRouterTest {
                                 .toInstance(msgRetryIntervalMs);
                 bind(Integer.class).annotatedWith(Names.named(ConfigurableMessagingSettings.PROPERTY_MESSAGING_MAXIMUM_PARALLEL_SENDS))
                                    .toInstance(maximumParallelSends);
+                bind(Long.class).annotatedWith(Names.named(ConfigurableMessagingSettings.PROPERTY_ROUTING_TABLE_GRACE_PERIOD_MS))
+                                .toInstance(routingTableGracePeriodMs);
+                bind(Long.class).annotatedWith(Names.named(ConfigurableMessagingSettings.PROPERTY_ROUTING_TABLE_CLEANUP_INTERVAL_MS))
+                                .toInstance(routingTableCleanupIntervalMs);
+
                 bind(Integer.class).annotatedWith(Names.named(ConfigurableMessagingSettings.PROPERTY_MAX_MESSAGES_INQUEUE))
                                    .toInstance(maxMessagesInQueue);
                 bindConstant().annotatedWith(Names.named(ClusterControllerRuntimeModule.PROPERTY_ACCESSCONTROL_ENABLE))
@@ -158,7 +165,9 @@ public class CcMessageRouterTest {
                                                                                  new HashSet<JoynrMessageProcessor>());
 
                 final boolean isGloballyVisible = true; // toParticipantId is globally visible
-                routingTable.put(toParticipantId, channelAddress, isGloballyVisible);
+                final long expiryDateMs = Long.MAX_VALUE;
+                final boolean isSticky = true;
+                routingTable.put(toParticipantId, channelAddress, isGloballyVisible, expiryDateMs, isSticky);
 
                 Request request = new Request("noMethod", new Object[]{}, new String[]{}, "requestReplyId");