+++ java/messaging/messaging-common/src/main/java/io/joynr/messaging/routing/RoutingTable.java
@@ -24,7 +24,7 @@ import joynr.system.RoutingTypes.Address;
 public interface RoutingTable {
     Address get(String participantId);
 
-    Address put(String participantId, Address address, boolean isGloballyVisible);
+    Address put(String participantId, Address address, boolean isGloballyVisible, long expiryDateMs, boolean isSticky);
 
     boolean containsKey(String participantId);
 
@@ -37,6 +37,16 @@ public interface RoutingTable {
      */
     boolean getIsGloballyVisible(String participantId);
 
+    /**
+     * Sets the isSticky attribute of the Routing Entry for the participantId.
+     * If true, the routing entry will not be get purged from routing table
+     * by the cleanup thread.
+     * @param participantId
+     * @param isSticky
+     * @throws JoynrRuntimeException if no entry exists for the given participantId
+     */
+    void setIsSticky(String participantId, boolean isSticky);
+
     void remove(String participantId);
 
     /**
@@ -46,4 +56,9 @@ public interface RoutingTable {
      *            the address operation to perform.
      */
     void apply(AddressOperation addressOperation);
+
+    /**
+     * Purge all expired routing entries from the table
+     */
+    void purge();
 }