+++ java/messaging/messaging-common/src/main/java/io/joynr/messaging/routing/RoutingTableImpl.java
@@ -21,6 +21,7 @@ package io.joynr.messaging.routing;
 
 import java.util.Map.Entry;
 import java.util.concurrent.ConcurrentMap;
+import java.util.Iterator;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -37,9 +38,11 @@ public class RoutingTableImpl implements RoutingTable {
     private static final Logger logger = LoggerFactory.getLogger(RoutingTableImpl.class);
 
     private static class RoutingEntry {
-        RoutingEntry(Address address, boolean isGloballyVisible) {
+        RoutingEntry(Address address, boolean isGloballyVisible, long expiryDateMs, boolean isSticky) {
             setAddress(address);
             setIsGloballyVisible(isGloballyVisible);
+            this.expiryDateMs = expiryDateMs;
+            this.isSticky = isSticky;
         }
 
         public Address getAddress() {
@@ -50,6 +53,14 @@ public class RoutingTableImpl implements RoutingTable {
             return isGloballyVisible;
         }
 
+        public long getExpiryDateMs() {
+            return expiryDateMs;
+        }
+
+        public boolean getIsSticky() {
+            return isSticky;
+        }
+
         public void setAddress(Address address) {
             this.address = address;
         }
@@ -58,8 +69,18 @@ public class RoutingTableImpl implements RoutingTable {
             this.isGloballyVisible = isGloballyVisible;
         }
 
+        public void setExpiryDateMs(long expiryDateMs) {
+            this.expiryDateMs = expiryDateMs;
+        }
+
+        public void setIsSticky(boolean isSticky) {
+            this.isSticky = isSticky;
+        }
+
         private Address address;
         private boolean isGloballyVisible;
+        private long expiryDateMs;
+        private boolean isSticky;
     }
 
     private ConcurrentMap<String, RoutingEntry> hashMap = Maps.newConcurrentMap();
@@ -86,6 +107,10 @@ public class RoutingTableImpl implements RoutingTable {
                        .append(eachEntry.getValue().address)
                        .append("\t-\t")
                        .append(eachEntry.getValue().isGloballyVisible)
+                       .append("\t-\t")
+                       .append(eachEntry.getValue().expiryDateMs)
+                       .append("\t-\t")
+                       .append(eachEntry.getValue().isSticky)
                        .append("\n");
             }
             logger.trace(message.toString());
@@ -93,12 +118,18 @@ public class RoutingTableImpl implements RoutingTable {
     }
 
     @Override
-    public Address put(String participantId, Address address, boolean isGloballyVisible) {
-        logger.trace("entering put(participantId={}, address={}, isGloballyVisible={})",
+    public Address put(String participantId,
+                       Address address,
+                       boolean isGloballyVisible,
+                       long expiryDateMs,
+                       boolean sticky) {
+        logger.trace("entering put(participantId={}, address={}, isGloballyVisible={}, expiryDateMs={}, sticky={})",
                      participantId,
                      address,
-                     isGloballyVisible);
-        RoutingEntry routingEntry = new RoutingEntry(address, isGloballyVisible);
+                     isGloballyVisible,
+                     expiryDateMs,
+                     sticky);
+        RoutingEntry routingEntry = new RoutingEntry(address, isGloballyVisible, expiryDateMs, sticky);
         RoutingEntry result = hashMap.putIfAbsent(participantId, routingEntry);
         // NOTE: ConcurrentMap cannot contain null values, this means if result is not null the new
         //       address was not added to the routing table
@@ -106,20 +137,36 @@ public class RoutingTableImpl implements RoutingTable {
         // Otherwise it returns the old mapped V and no insertion to the Routing table takes place
         if (result != null) {
             if (!address.equals(result.getAddress()) || result.getIsGloballyVisible() != isGloballyVisible) {
-                logger.warn("unable to update(participantId={}, address={}, isGloballyVisible={}) into routing table,"
+                logger.warn("unable to update(participantId={}, address={}, isGloballyVisible={}, expiryDateMs={}, sticky={}) into routing table,"
                                     + " since the participant ID is already associated with routing entry address={}, isGloballyVisible={}",
                             participantId,
                             address,
                             isGloballyVisible,
                             address,
-                            isGloballyVisible);
+                            isGloballyVisible,
+                            expiryDateMs,
+                            sticky);
+            } else {
+                // address and isGloballyVisible are identical
+
+                // extend lifetime, if required
+                if (result.getExpiryDateMs() < expiryDateMs) {
+                    result.setExpiryDateMs(expiryDateMs);
+                }
+                // make entry sticky, if required
+                // if entry already was sticky, and new entry is not, keep the sticky attribute
+                if (sticky && !result.getIsSticky()) {
+                    result.setIsSticky(true);
+                }
             }
             return result.getAddress();
         } else {
-            logger.trace("put(participantId={}, address={}, isGloballyVisible={}) successfully into routing table",
+            logger.trace("put(participantId={}, address={}, isGloballyVisible={}, expiryDateMs={}, sticky={}) successfully into routing table",
                          participantId,
                          address,
-                         isGloballyVisible);
+                         isGloballyVisible,
+                         expiryDateMs,
+                         sticky);
             return null;
         }
     }
@@ -138,13 +185,31 @@ public class RoutingTableImpl implements RoutingTable {
     public boolean getIsGloballyVisible(String participantId) {
         RoutingEntry routingEntry = hashMap.get(participantId);
         if (routingEntry == null) {
-            throw new JoynrRuntimeException("participateId doesn't exist in the routing table");
+            throw new JoynrRuntimeException("participantId doesn't exist in the routing table");
         }
         return routingEntry.getIsGloballyVisible();
     }
 
     @Override
+    public void setIsSticky(String participantId, boolean isSticky) {
+        RoutingEntry routingEntry = hashMap.get(participantId);
+        if (routingEntry == null) {
+            throw new JoynrRuntimeException("participantId doesn't exist in the routing table");
+        }
+        routingEntry.setIsSticky(isSticky);
+    }
+
+    @Override
     public void remove(String participantId) {
+        RoutingEntry routingEntry = hashMap.get(participantId);
+        if (routingEntry != null) {
+            logger.trace("removing(participantId={}, address={}, isGloballyVisible={}, expiryDateMs={}, sticky={}) from routing table",
+                         participantId,
+                         routingEntry.getAddress(),
+                         routingEntry.getIsGloballyVisible(),
+                         routingEntry.getExpiryDateMs(),
+                         routingEntry.getIsSticky());
+        }
         hashMap.remove(participantId);
     }
 
@@ -158,4 +223,32 @@ public class RoutingTableImpl implements RoutingTable {
         }
     }
 
+    public void purge() {
+        logger.trace("purge: begin");
+        Iterator<Entry<String, RoutingEntry>> it = hashMap.entrySet().iterator();
+        long currentTimeMillis = System.currentTimeMillis();
+        while (it.hasNext()) {
+            Entry<String, RoutingEntry> e = it.next();
+            if (logger.isTraceEnabled()) {
+                logger.trace("check: participantId = {}, sticky = {}, expiryDateMs = {}, now = {}",
+                             e.getKey(),
+                             e.getValue().getIsSticky(),
+                             e.getValue().getExpiryDateMs(),
+                             currentTimeMillis);
+            }
+
+            if (!e.getValue().getIsSticky() && e.getValue().expiryDateMs < currentTimeMillis) {
+                if (logger.isTraceEnabled()) {
+                    logger.trace("purging(participantId={}, address={}, isGloballyVisible={}, expiryDateMs={}, sticky={}) from routing table",
+                                 e.getKey(),
+                                 e.getValue().getAddress(),
+                                 e.getValue().getIsGloballyVisible(),
+                                 e.getValue().getExpiryDateMs(),
+                                 e.getValue().getIsSticky());
+                }
+                it.remove();
+            }
+        }
+        logger.trace("purge: end");
+    }
 }